CollideStrategyPattern에서 충돌전략을 CollideStrategy 클래스로 분리하고  CollideHandler로 rename
